pyproject.toml
[project]
name = "rocktimes"
version = "0.1.0"
description = "Rocktimes Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.3",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "strawberry-graphql[debug-server,fastapi]>=0.247.0",
]
